make suggested changes.

+++ ex_feels/lib/ex_feels/stat.ex
@@ -10,7 +10,7 @@ defmodule ExFeels.Stat do
     field :classifier, :string
     field :mean, :float
     field :std, :float
-    field :time, :float
+    field :time, :integer
   end
 
   def all() do
